requirements.txt
Flask==1.0.2
gunicorn==19.9.0
google-cloud-datastore==1.7.0
